Cleaned up code to register both players
With reservations. The label and ordering isn't correct on ENet.

Apparently I was using an add_player function for purposes which might
be better suited to a register_player rpc call. Lastly, adding the
player after joining the lobby requires a local set to players, which is
interesting, as I feel like "call_local" should prevent that.

Worthy of further consideration on condensing.