Removed peer_id naming requirement
In the previous version, the name of the player was always set to
peer_id, which then needed to be converted back every time we
wanted to check it, and then converted to an int. No good for lots of
different reasons.

All we really needed it for was setting authority; which is now done
with an RPC call.